#include<stdio.h> int main() {     int n;     scanf("%d",&n);     int a[n],i,s=0,c=0;     for(i=0;i<n;i++)     {         scanf("%d",&a[i]);     }     for(i=0;i<n;i++)     {         s=s+a[i];     }     int avg= s/n;     for(i=0;i<n;i++)     {         if(avg>=a[i])         {             c++;         }     }     printf("%d",c); }

+++ elements_smaller_than_average.c
@@ -12,12 +12,12 @@ int main()
     {
         s=s+a[i];
     }
-    float avg=(float)s/n;
+    int avg= s/n;
     for(i=0;i<n;i++)
     {
         if(avg>=a[i])
         {
-           c++;
+            c++;
         }
     }
     printf("%d",c);
